Clean up in client-core/common/component
