+++ packages/client-core/src/common/components/ErrorBoundary.tsx
@@ -1,9 +1,37 @@
 import React from 'react'
 
-import Catch from './functionalErrorBoundary'
-
 type Props = {
   children: React.ReactNode
+}
+
+type ErrorHandler = (error: Error, info: React.ErrorInfo) => void
+type ErrorHandlingComponent<Props> = (props: Props, error?: Error) => React.ReactNode
+
+type ErrorState = { error?: Error }
+
+function Catch<Props extends {}>(
+  component: ErrorHandlingComponent<Props>,
+  errorHandler?: ErrorHandler
+): React.ComponentType<Props> {
+  return class extends React.Component<Props, ErrorState> {
+    state: ErrorState = {
+      error: undefined
+    }
+
+    static getDerivedStateFromError(error: Error) {
+      return { error }
+    }
+
+    componentDidCatch(error: Error, info: React.ErrorInfo) {
+      if (errorHandler) {
+        errorHandler(error, info)
+      }
+    }
+
+    render() {
+      return component(this.props, this.state.error)
+    }
+  }
 }
 
 const ErrorBoundary = Catch(function error(props: Props, error?: Error) {
