+++ packages/client/src/pages/_app.tsx
@@ -7,7 +7,7 @@ import {
   ClientSettingService,
   useClientSettingState
 } from '@xrengine/client-core/src/admin/services/Setting/ClientSettingService'
-import { initGA, logPageView } from '@xrengine/client-core/src/common/components/analytics'
+import { initGA, logPageView } from '@xrengine/client-core/src/common/analytics'
 import MetaTags from '@xrengine/client-core/src/common/components/MetaTags'
 import { defaultAction } from '@xrengine/client-core/src/common/components/NotificationActions'
 import { ProjectService, useProjectState } from '@xrengine/client-core/src/common/services/ProjectService'
